refactor(object): Agree with the tileset on what a tile id is
TileSet::get already took a tile id as unsigned and add_tile took it signed,
though both read the same uint32_t the tileset file wrote, so an id past the
top of an int reached the setter as a negative and indexed behind the array.
Giving add_tile the type its callers already hold removes that, and the id can
no longer be a number the vector cannot hold.

The tilemap's own indexes stay signed, since they are checked against the map's
bounds before use and the parser rejects a negative width or height outright.
They cross to the container's type at the point of use instead.

+++ src/object/path_walker.cpp
@@ -131,7 +131,7 @@ PathWalker::advance_node()
   if (static_cast<int>(current_node_nr) == stop_at_node_nr) running = false;
 
   if (path->mode == Path::UNORDERED) {
-    next_node_nr = gameRandom.rand( path->nodes.size() );
+    next_node_nr = static_cast<size_t>(gameRandom.rand(static_cast<int>(path->nodes.size())));
     return;
   }
